+++ src/filters.tsx
@@ -18,6 +18,7 @@ addFilter( 'cbtj.innerBlocks', 'cbtj', ( innerBlocks, block ) => {
 	switch ( block ) {
 		case 'core/list':
 		case 'core/details':
+		case 'core/pullquote':
 			blocks = innerBlocks.map( ( { name, attributes } ) =>
 				createBlock( name, { ...JSON.parse( attributes ) } )
 			);