+++ tests/unit/php/Services/BlocksTest.php
@@ -8,6 +8,7 @@ use Badasswp\WPMockTC\WPMockTestCase;
 use ConvertBlocksToJSON\Abstracts\Block;
 use ConvertBlocksToJSON\Services\Blocks;
 
+use ConvertBlocksToJSON\Blocks\Pullquote;
 use ConvertBlocksToJSON\Blocks\Details;
 use ConvertBlocksToJSON\Blocks\Heading;
 use ConvertBlocksToJSON\Blocks\Lists;
@@ -37,6 +38,7 @@ class BlocksTest extends WPMockTestCase {
 	public function test_class_properties_are_defined_by_default() {
 		$this->assertSame(
 			[
+				Pullquote::class,
 				Details::class,
 				Heading::class,
 				Lists::class,
@@ -80,6 +82,7 @@ class BlocksTest extends WPMockTestCase {
 		WP_Mock::expectFilter(
 			'cbtj_blocks',
 			[
+				Pullquote::class,
 				Details::class,
 				Heading::class,
 				Lists::class,